Feat : 기록 조회 및 댓글 조회 기능 구현중...
